redeloy

+++ components/Orders/OrderModal.tsx
@@ -42,7 +42,11 @@ const OrderModal = ({
       );
     });
   }, [codebill, id, shipping_fee, isDelete]);
-  const handleDeleteOrderDetail = (orderDetailId: number) => {
+  const handleDeleteOrderDetail = (orderDetailId: number, status: string) => {
+    if (orderDetailId && status == "s4") {
+      message.success("Bạn không thể xóa đơn hàng đã giao thành công");
+      return;
+    }
     orderDetailId &&
       OrderApi.DeleteOrderDetail(orderDetailId)
         .then((res: any) => {
@@ -135,7 +139,9 @@ const OrderModal = ({
                     <td className="border-b border-[#eee] py-5 px-4 dark:border-strokedark">
                       <Popconfirm
                         title="Bạn có chắc muốn xóa không?"
-                        onConfirm={() => handleDeleteOrderDetail(order.id)}
+                        onConfirm={() =>
+                          handleDeleteOrderDetail(order.id, selected?.status)
+                        }
                         okText="Xác nhận"
                         cancelText="Hủy"
                         okType="danger"
@@ -172,7 +178,7 @@ const OrderModal = ({
               pageSize={itemPerPage}
               current={pageCurrentModal}
               onChange={(page) => setPageCurrentModal(page)}
-                showSizeChanger={false}
+              showSizeChanger={false}
             />
           </div>
         </div>
